big_data_mongodb/requirement.txt
streamlit
pymongo
bson
pandas
cryptography
python-dotenv